Added integration test for database connection
